configLoader.go
package main

import (
	"errors"
	"fmt"
	"github.com/joho/godotenv"
	"os"
	"strconv"
	"time"
)

type Config struct {
	redisDsn      string
	kafkaHost     string
	kafkaTimeout  time.Duration
	kafkaAttempts int
}

func loadConfig(envFilename string) (Config, error) {
	if envFilename != "" {
		err := godotenv.Load(envFilename)
		if err != nil {
			return Config{}, errors.New(fmt.Sprintf("Error loading %s file: %s", envFilename, err))
		}
	}

	kafkaTimeout, err := strconv.Atoi(os.Getenv("KAFKA_TIMEOUT"))
	if kafkaTimeout == 0 || err != nil {
		kafkaTimeout = 10
	}

	kafkaAttempts, err := strconv.Atoi(os.Getenv("KAFKA_ATTEMPTS"))
	if kafkaAttempts == 0 || err != nil {
		kafkaAttempts = 0
	}

	config := Config{
		redisDsn:      os.Getenv("REDIS_DSN"),
		kafkaHost:     os.Getenv("KAFKA_HOST"),
		kafkaTimeout:  time.Second * time.Duration(kafkaTimeout),
		kafkaAttempts: kafkaAttempts,
	}

	if config.kafkaHost == "" {
		return Config{}, errors.New("empty KAFKA_HOST")
	}

	return config, nil
}
